+++ calendar/CopticCalendar.hpp
@@ -24,14 +24,14 @@ namespace EpsilonDelta
 class CopticCalendar
 {
 public:
-    static void JulianDayToDMY( int julianDay,
-                                int * pDay, int * pMonth, int * pYear );
-    static int DMYToJulianDay( int day, int month, int year );
-    static int MonthsInYear( int year );
-    static int DaysInMonth( int month, int year );
-    static const std::string & MonthName( int month, int year );
+    static void JulianDayToDMY( long julianDay,
+                                int * pDay, int * pMonth, long * pYear );
+    static long DMYToJulianDay( int day, int month, long year );
+    static int MonthsInYear( long year );
+    static int DaysInMonth( int month, long year );
+    static const std::string & MonthName( int month, long year );
     static const std::string & MonthName( int month );
-    static bool IsLeapYear( int year );
+    static bool IsLeapYear( long year );
 
     enum EMonth
     { Thoout = 1, Paope, Athor, Koiak, Tobe, Meshir, Paremotep,
@@ -44,7 +44,7 @@ public:
 
 inline
 int
-CopticCalendar::MonthsInYear( int /*year*/ )
+CopticCalendar::MonthsInYear( long /*year*/ )
 {
     return 13;  //including Epagomene.
 }
